moving Report and Page classes to separate files
