+++ trustar/__init__.py
@@ -1,4 +1,6 @@
 # from TruStar import TruStar
 from __future__ import absolute_import
 
-from .TruStar import *
+from .trustar import TruStar
+from .report import Report, DISTRIBUTION_TYPE_COMMUNITY, DISTRIBUTION_TYPE_ENCLAVE
+from .page import Page
