+++ trustar/trustar.py
@@ -1,31 +1,30 @@
 from __future__ import print_function
 
-import json
-import time
-from datetime import datetime
+from .report import Report, DISTRIBUTION_TYPE_ENCLAVE
+from .page import Page
+from .utils import normalize_timestamp
 
+import logging
+import json
+from datetime import datetime
 import configparser
-import dateutil.parser
-import pytz
-import math
 import requests
 import requests.auth
+from requests import HTTPError
 from builtins import object
 from future import standard_library
-from tzlocal import get_localzone
-from requests import HTTPError\
 
 standard_library.install_aliases()
 
-CLIENT_VERSION = "0.3.0"
+logger = logging.getLogger(__name__)
 
-DISTRIBUTION_TYPE_ENCLAVE = "ENCLAVE"
-DISTRIBUTION_TYPE_COMMUNITY = "COMMUNITY"
+
+CLIENT_VERSION = "0.3.0"
 
 
 class TruStar(object):
     """
-    Main class you to instantiate the TruStar API
+    This class is used to interact with the TruStar API.
     """
 
     REQUIRED_KEYS = ['auth', 'base', 'api_key', 'api_secret']
@@ -37,6 +36,13 @@ class TruStar(object):
     }
 
     def __init__(self, config_file="../examples/trustar.conf", config_role="trustar", config=None):
+        """
+        Constructs and configures the instance.  Attempts to use 'config' to configure.  If it is None,
+        then attempts to use 'config_file' instead.
+        :param config_file: Path to configuration file.
+        :param config_role: The section in the configuration file to use.
+        :param config: A dictionary of configuration options.
+        """
 
         # attempt to use configuration file if one exists
         if config is None:
@@ -90,49 +96,9 @@ class TruStar(object):
         self.verify = config['verify']
         self.enclave_ids = config['enclave_ids']
 
-
     @staticmethod
     def normalize_timestamp(date_time):
-        """
-        Attempt to convert a string timestamp in to a TruSTAR compatible format for submission.
-        Will return current time with UTC time zone if None
-        :param date_time: int that is epoch time, or string/datetime object containing date, time, and ideally timezone
-        examples of supported timestamp formats: 1487890914, 1487890914000, "2017-02-23T23:01:54", "2017-02-23T23:01:54+0000"
-        """
-        datetime_dt = datetime.now()
-
-        # get current time in seconds-since-epoch
-        current_time = int(time.time())
-
-        try:
-            # identify type of timestamp and convert to datetime object
-            if isinstance(date_time, int):
-
-                # if timestamp has more than 10 digits, it is in ms
-                if date_time > 9999999999:
-                    date_time /= 1000
-
-                # if timestamp is incorrectly forward dated, set to current time
-                if date_time > current_time:
-                    date_time = current_time
-                datetime_dt = datetime.fromtimestamp(date_time)
-            elif isinstance(date_time, str):
-                datetime_dt = dateutil.parser.parse(date_time)
-            elif isinstance(date_time, datetime):
-                datetime_dt = date_time
-
-        # if timestamp is none of the formats above, error message is printed and timestamp is set to current time by default
-        except Exception as e:
-            print(e)
-            datetime_dt = datetime.now()
-
-        # if timestamp is timezone naive, add timezone
-        if not datetime_dt.tzinfo:
-            # add system timezone and convert to UTC
-            datetime_dt = get_localzone().localize(datetime_dt).astimezone(pytz.utc)
-
-        # converts datetime to iso8601
-        return datetime_dt.isoformat()
+        return normalize_timestamp(date_time)
 
     def __get_token(self):
         """
@@ -157,9 +123,7 @@ class TruStar(object):
         :param is_json: Whether the request body is a json.
         :return: The headers dictionary.
         """
-        headers = {
-            "Authorization": "Bearer " + self.__get_token(),
-        }
+        headers = {"Authorization": "Bearer " + self.__get_token()}
 
         if self.client_type is not None:
             headers["Client-Type"] = self.client_type
@@ -245,7 +209,6 @@ class TruStar(object):
         """
         Build direct URL to report from its ID
         :param report_id: Incident Report (IR) ID, e.g., as returned from `submit_report`
-        :param kwargs: Any extra keyword arguments.  These will be forwarded to requests.request.
         :return URL
         """
 
@@ -300,8 +263,8 @@ class TruStar(object):
         (by default all accessible reports are returned). Possible values are: 'COMMUNITY' and 'ENCLAVE'
         :param enclave_ids: Optional comma separated list of enclave ids, restrict reports to specific enclaves
         (by default reports from all enclaves are returned)
-        :param from_time: Optional start of time window (Unix timestamp - seconds since epoch)
-        :param to_time: Optional end of time window (Unix timestamp - seconds since epoch)
+        :param from_time: Optional start of time window (Unix timestamp - milliseconds since epoch)
+        :param to_time: Optional end of time window (Unix timestamp - milliseconds since epoch)
         :param page_number: The page number to get.
         :param page_size: The size of the page to be returned.
         :param kwargs: Any extra keyword arguments.  These will be forwarded to requests.request.
@@ -320,10 +283,12 @@ class TruStar(object):
             'pageSize': page_size
         }
         resp = self.__get("reports", params=params, **kwargs)
+        body = resp.json()
 
-        body = json.loads(resp.content.decode('utf8'))
+        # replace each dict in 'items' with a Report object
         body['items'] = [Report.from_dict(report) for report in body['items']]
 
+        # create a Page object from the dict
         return Page.from_dict(body)
 
 
@@ -339,12 +304,18 @@ class TruStar(object):
         :param external_url: external url of report, optional and is associated with the original source of this report
         :param time_began: time report began
         :param enclave: boolean - whether or not to submit report to user's enclaves (see 'enclave_ids' config property)
+        :param enclave_ids: the IDs of the enclaves to submit the report to
+        :param report: a Report object.  If present, other parameters will be ignored.
         :param kwargs: Any extra keyword arguments.  These will be forwarded to requests.request.
         """
 
+        # if no Report object was passed, construct one from the other parameters
         if report is None:
+
+            # use configured enclave_ids by default
             if enclave_ids is None:
                 enclave_ids = self.enclave_ids
+
             report = Report(title=title,
                             body=report_body,
                             time_began=time_began,
@@ -370,25 +341,32 @@ class TruStar(object):
         :param title: new title for report
         :param report_body: new body for report
         :param time_began: new time_began for report
+        :param external_url: external url of report, optional and is associated with the original source of this report
         :param distribution: new distribution type for report
         :param enclave_ids: new list of enclave ids that the report will belong to (python list or comma-separated list)
-        :param external_url: external url of report, optional and is associated with the original source of this report
+        :param report: a Report object.  If present, other parameters will be ignored.
         :param kwargs: Any extra keyword arguments.  These will be forwarded to requests.request.
         """
 
+        # if no Report object was passed, construct one from the other parameters
         if report is None:
+
+            # use configured enclave_ids by default
             if enclave_ids is None:
                 enclave_ids = self.enclave_ids
+
             report = Report(id=report_id,
                             title=title,
                             body=report_body,
                             time_began=time_began,
                             external_url=external_url,
-                            is_enclave=distribution is None or distribution.upper() == Report.DISTRIBUTION_TYPE_ENCLAVE,
+                            is_enclave=distribution is None or distribution.upper() == DISTRIBUTION_TYPE_ENCLAVE,
                             enclave_ids=enclave_ids)
 
+        # make id_type default to "internal"
         id_type = id_type or Report.ID_TYPE_INTERNAL
 
+        # determine which ID to use based on id_type
         if id_type.upper() == Report.ID_TYPE_EXTERNAL:
             report_id = report.external_id
         else:
@@ -402,7 +380,7 @@ class TruStar(object):
         }
 
         resp = self.__put("report/%s" % report_id, data=json.dumps(payload), params=params, **kwargs)
-        return json.loads(resp.content.decode('utf8'))
+        return resp.json()
 
     def delete_report(self, report_id, id_type=None, **kwargs):
         """
@@ -422,60 +400,48 @@ class TruStar(object):
         :param indicators: The list of indicators to retrieve correlated reports for.
         :param kwargs: Any extra keyword arguments.  These will be forwarded to requests.request.
         """
-        payload = {'indicators': indicators}
-        resp = self.__get("reports/correlate", params=payload, **kwargs)
-        return json.loads(resp.content.decode('utf8'))
+        params = {'indicators': indicators}
+        resp = self.__get("reports/correlate", params=params, **kwargs)
+        return resp.json()
 
 
     ###########################
     ### Indicator Endpoints ###
     ###########################
 
-    # def query_latest_indicators(self, source, indicator_types, limit, interval_size, **kwargs):
-    #     """
-    #     Finds all latest indicators
-    #     :param source: source of the indicators which can either be INCIDENT_REPORT or OSINT
-    #     :param indicator_types: a list of indicators or a string equal to "ALL" to query all indicator types extracted
-    #     by TruSTAR
-    #     :param limit: limit on the number of indicators. Max is set to 5000
-    #     :param interval_size: time interval on returned indicators. Max is set to 24 hours
-    #     :param kwargs: Any extra keyword arguments.  These will be forwarded to requests.request.
-    #     :return json response of the result
-    #     """
-    #     payload = {'source': source, 'types': indicator_types, 'limit': limit, 'intervalSize': interval_size}
-    #     resp = self.__get("indicators/latest", params=payload, **kwargs)
-    #     return json.loads(resp.content.decode('utf8'))
-
     def get_community_trends(self, indicator_type=None, from_time=None, to_time=None, page_size=None, page_number=None, **kwargs):
         """
         Find community trending indicators.
-        :param indicator_type: the type of indicators.  3 types are supported: "malware", "cve" (vulnerabilities),
-        "other" (all
-        other types of indicators)
-        :param from_time: Optional start of time window (Unix timestamp - seconds since epoch)
-        :param to_time: Optional end of time window (Unix timestamp - seconds since epoch)
+        :param indicator_type: the type of indicators.  If None, will get all types of indicators except for MALWARE and CVEs.
+        :param from_time: Optional start of time window (Unix timestamp - milliseconds since epoch)
+        :param to_time: Optional end of time window (Unix timestamp - milliseconds since epoch)
         :param page_size: # of results on returned page
-        :param start_page: page to start returning results on
+        :param page_number: page to start returning results on
         :param kwargs: Any extra keyword arguments.  These will be forwarded to requests.request.
         :return: json response of the result
         """
 
-        payload = {
+        params = {
             'type': indicator_type,
             'from': from_time,
             'to': to_time,
             'pageSize': page_size,
             'pageNumber': page_number
         }
-        resp = self.__get("indicators/community-trending", params=payload, **kwargs)
-        return Page.from_dict(json.loads(resp.content.decode('utf8')))
+        resp = self.__get("indicators/community-trending", params=params, **kwargs)
+        return Page.from_dict(resp.json())
 
-    def get_related_indicators(self, indicators=None, sources=None, page_number=None, page_size=None, **kwargs):
+    def get_related_indicators(self, indicators=None, sources=None, page_size=None, page_number=None, **kwargs):
         """
-         Finds all reports that contain the indicators and returns correlated indicators from those reports.
-         :param indicators: list of indicators to search for
-         :param kwargs: Any extra keyword arguments.  These will be forwarded to requests.request.
-         """
+        Finds all reports that contain the indicators and returns correlated indicators from those reports.
+        :param indicators: list of indicators to search for
+        :param sources: list of sources to search.  Options are: INCIDENT_REPORT, EXTERNAL_INTELLIGENCE, and ORION_FEED.
+        :param page_size: # of results on returned page
+        :param page_number: page to start returning results on
+        :param kwargs: Any extra keyword arguments.  These will be forwarded to requests.request.
+        :return: json response of the result
+        """
+
         params = {
             'indicators': indicators,
             'sources': sources,
@@ -483,20 +449,22 @@ class TruStar(object):
             'pageSize': page_size
         }
         resp = self.__get("indicators/related", params=params, **kwargs)
-        return Page.from_dict(json.loads(resp.content.decode('utf8')))
+        return Page.from_dict(resp.json())
 
     def get_related_external_indicators(self, indicators=None, sources=None, **kwargs):
         """
-         Finds all reports that contain the indicators and returns correlated indicators from those reports.
-         :param indicators: list of indicators to search for
-         :param kwargs: Any extra keyword arguments.  These will be forwarded to requests.request.
-         """
+        Finds all reports that contain the indicators and returns correlated indicators from those reports.
+        :param indicators: list of indicators to search for
+        :param sources: list of sources to search
+        :param kwargs: Any extra keyword arguments.  These will be forwarded to requests.request.
+        """
+
         params = {
             'indicators': indicators,
             'sources': sources
         }
         resp = self.__get("indicators/external/related", params=params, **kwargs)
-        return json.loads(resp.content.decode('utf8'))
+        return resp.json()
 
 
     #####################
@@ -513,7 +481,7 @@ class TruStar(object):
         """
         params = {'idType': id_type}
         resp = self.__get("reports/%s/enclave-tags" % report_id, params=params, **kwargs)
-        return json.loads(resp.content.decode('utf8'))
+        return resp.json()
 
     def add_enclave_tag(self, report_id, name, enclave_id, id_type=None, **kwargs):
         """
@@ -531,7 +499,7 @@ class TruStar(object):
             'enclaveId': enclave_id
         }
         resp = self.__post("reports/%s/enclave-tags" % report_id, params=params, **kwargs)
-        return json.loads(resp.content.decode('utf8'))
+        return resp.json()
 
     def delete_enclave_tag(self, report_id, name, enclave_id, id_type=None, **kwargs):
         """
@@ -557,7 +525,14 @@ class TruStar(object):
     #################
 
     @staticmethod
-    def get_page_iterator(func, start_page=0, page_size=None):
+    def get_page_generator(func, start_page=0, page_size=None):
+        """
+        Gets a generator for retrieving pages from a paginated endpoint.
+        :param func: Should take parameters 'page_number' and 'page_size' and return the corresponding Page object.
+        :param start_page: The page to start on.
+        :param page_size: The size of each page.
+        :return: The generator.
+        """
         page_number = start_page
         more_pages = True
         while more_pages:
@@ -567,162 +542,86 @@ class TruStar(object):
             page_number += 1
 
     @classmethod
-    def get_iterator(cls, func=None, page_iterator=None):
+    def get_generator(cls, func=None, page_iterator=None):
+        """
+        Gets a generator for retrieving all results from a paginated endpoint.  Pass exactly one of 'page_iterator'
+        or 'func'.
+        :param func: Should take parameters 'page_number' and 'page_size' and return the corresponding Page object.
+        If page_iterator is None, this will be used to create one.
+        :param page_iterator: A page_iterator to be used to generate each successive page.
+        :return: The generator.
+        """
 
+        # if page_iterator is None, use func to create one
         if page_iterator is None:
             if func is None:
                 raise Exception("To use 'get_iterator', must provide either a page iterator or a method.")
             else:
-                page_iterator = cls.get_page_iterator(func)
+                page_iterator = cls.get_page_generator(func)
 
         for page in page_iterator:
             for item in page.items:
                 yield item
 
     def get_report_page_iterator(self, start_page=0, page_size=None, **kwargs):
+        """
+        Creates a generator from the 'get_reports' method that returns each successive page.
+        :param start_page: The page to start on.
+        :param page_size: The size of each page.
+        :param kwargs: Any extra keyword arguments.  These will be forwarded to the 'get_reports' method.
+        :return: The generator.
+        """
         def func(page_number, page_size):
             return self.get_reports(page_number=page_number, page_size=page_size, **kwargs)
 
-        return self.get_page_iterator(func, start_page, page_size)
+        return self.get_page_generator(func, start_page, page_size)
 
     def get_report_iterator(self, **kwargs):
-        return self.get_iterator(page_iterator=self.get_report_page_iterator(**kwargs))
+        """
+        Creates a generator from the 'get_reports' method that returns each successive report.
+        :param kwargs: Any extra keyword arguments.  These will be forwarded to the 'get_reports' method.
+        :return: The generator.
+        """
+        return self.get_generator(page_iterator=self.get_report_page_iterator(**kwargs))
 
     def get_community_trends_page_iterator(self, start_page=0, page_size=None, **kwargs):
+        """
+        Creates a generator from the 'get_community_trends' method that returns each successive page.
+        :param start_page: The page to start on.
+        :param page_size: The size of each page.
+        :param kwargs: Any extra keyword arguments.  These will be forwarded to the 'get_community_trends' method.
+        :return: The generator.
+        """
         def func(page_number, page_size):
             return self.get_community_trends(page_number=page_number, page_size=page_size, **kwargs)
 
-        return self.get_page_iterator(func, start_page, page_size)
+        return self.get_page_generator(func, start_page, page_size)
 
     def get_community_trends_iterator(self, **kwargs):
-        return self.get_iterator(page_iterator=self.get_community_trends_page_iterator(**kwargs))
+        """
+        Creates a generator from the 'get_community_trends_iterator' method that returns each successive report.
+        :param kwargs: Any extra keyword arguments.  These will be forwarded to the 'get_community_trends_iterator' method.
+        :return: The generator.
+        """
+        return self.get_generator(page_iterator=self.get_community_trends_page_iterator(**kwargs))
 
     def get_related_indicators_page_iterator(self, start_page=0, page_size=None, **kwargs):
+        """
+        Creates a generator from the 'get_related_indicators' method that returns each successive page.
+        :param start_page: The page to start on.
+        :param page_size: The size of each page.
+        :param kwargs: Any extra keyword arguments.  These will be forwarded to the 'get_related_indicators' method.
+        :return: The generator.
+        """
         def func(page_number, page_size):
             return self.get_related_indicators(page_number=page_number, page_size=page_size, **kwargs)
 
-        return self.get_page_iterator(func, start_page, page_size)
+        return self.get_page_generator(func, start_page, page_size)
 
     def get_related_indicators_iterator(self, **kwargs):
-        return self.get_iterator(page_iterator=self.get_related_indicators_page_iterator(**kwargs))
-
-
-class Report(object):
-
-    ID_TYPE_INTERNAL = "INTERNAL"
-    ID_TYPE_EXTERNAL = "EXTERNAL"
-
-    def __init__(self,
-                 id=None,
-                 title=None,
-                 body=None,
-                 time_began=None,
-                 external_id=None,
-                 external_url=None,
-                 is_enclave=True,
-                 enclave_ids=None):
-
-        if is_enclave:
-
-            # if string, convert comma-separated list into python list
-            if isinstance(enclave_ids, str) or isinstance(enclave_ids, unicode):
-                enclave_ids = [x.strip() for x in enclave_ids.split(',')]
-
-            # ensure is list
-            if not isinstance(enclave_ids, list):
-                raise ValueError("Enclave IDs must either be a list or a comma-separated string.")
-
-            # ensure non-empty
-            if len(enclave_ids) == 0:
-                raise ValueError("Enclave report must have one or more enclave IDs.")
-
-            # filter out None values
-            enclave_ids = [i for i in enclave_ids if i is not None]
-
-        self.id = id
-        self.title = title
-        self.body = body
-        self.time_began = time_began
-        self.external_id = external_id
-        self.external_url = external_url
-        self.is_enclave = is_enclave
-        self.enclave_ids = enclave_ids
-
-    def get_distribution_type(self):
-        if self.is_enclave:
-            return DISTRIBUTION_TYPE_ENCLAVE
-        else:
-            return DISTRIBUTION_TYPE_COMMUNITY
-
-    def to_dict(self, update=False):
-        result = {
-            'title': self.title,
-            'reportBody': self.body,
-            'timeBegan': TruStar.normalize_timestamp(self.time_began),
-            'externalUrl': self.external_url,
-            'distributionType': self.get_distribution_type()
-        }
-
-        if not update:
-            result['externalTrackingId'] = self.external_id
-
-        return result
-
-    @classmethod
-    def from_dict(cls, report):
-
-        is_enclave = report.get('distributionType')
-        if is_enclave is not None:
-            is_enclave = is_enclave.upper() != DISTRIBUTION_TYPE_COMMUNITY
-
-        return Report(
-            id=report.get('id'),
-            title=report.get('title'),
-            body=report.get('reportBody'),
-            time_began=report.get('timeBegan'),
-            external_url=report.get('externalUrl'),
-            is_enclave=is_enclave
-        )
-
-    def __str__(self):
-        return json.dumps(self.to_dict())
-
-
-class Page(object):
-
-    def __init__(self, items=None, page_number=None, page_size=None, total_elements=None):
-        self.items = items
-        self.page_number = page_number
-        self.page_size = page_size
-        self.total_elements = total_elements
-
-    def get_total_pages(self):
-        return math.ceil(self.total_elements / self.page_size)
-
-    def has_more_pages(self):
-        return self.page_number < self.get_total_pages()
-
-    @staticmethod
-    def from_dict(page):
-        return Page(items=page['items'],
-                    page_number=page['pageNumber'],
-                    page_size=page['pageSize'],
-                    total_elements=page['totalElements'])
-
-    def to_dict(self):
-        items = []
-        for item in self.items:
-            if hasattr(item, 'to_dict'):
-                items.append(item.to_dict())
-            else:
-                items.append(item)
-
-        return {
-            'items': items,
-            'pageNumber': self.page_number,
-            'pageSize': self.page_size,
-            'totalElements': self.total_elements
-        }
-
-    def __str__(self):
-        return json.dumps(self.to_dict())
+        """
+        Creates a generator from the 'get_generator' method that returns each successive report.
+        :param kwargs: Any extra keyword arguments.  These will be forwarded to the 'get_generator' method.
+        :return: The generator.
+        """
+        return self.get_generator(page_iterator=self.get_related_indicators_page_iterator(**kwargs))
